… projects/Amlogic/linux/linux.aarch64.conf