This adds the modules needed to use LUKS.

When `CONFIG_ARM64_CRYPTO` is defined, the macro `module_cpu_feature_match()` is used to check that the CPU supports the crypto extension.  Unfortunately, that macro won't compile and gives the error:

```
In file included from arch/arm64/crypto/aes-glue.c:16:0:
include/linux/module.h:87:32: error: storage size of '__mod_cpu_device_table' isn't known
 extern const struct gtype##_id __mod_##gtype##_table  \
                                ^
include/linux/module.h:145:3: note: in expansion of macro 'MODULE_GENERIC_TABLE'
   MODULE_GENERIC_TABLE(type##_device, name)
   ^~~~~~~~~~~~~~~~~~~~
include/linux/cpufeature.h:49:1: note: in expansion of macro 'MODULE_DEVICE_TABLE'
 MODULE_DEVICE_TABLE(cpu, cpu_feature_match_ ## x);  \
 ^~~~~~~~~~~~~~~~~~~
arch/arm64/crypto/aes-glue.c:442:1: note: in expansion of macro 'module_cpu_feature_match'
 module_cpu_feature_match(AES, aes_init);
 ^~~~~~~~~~~~~~~~~~~~~~~~
make[1]: *** [/src/CoreELEC-coreelec-9.0/build.CoreELEC-Amlogic.arm-9.0-devel/linux-f1bb00821950b0f5cab1b8a820c697ba9fbd3183/arch/arm64/crypto/Makefile:38: arch/arm64/crypto/aes-glue-ce.o] Error 1
```

I've included a patch to remove the macro. Since we're targeting a specific device and it will have the expected extensions it should be safe, but I'm open to suggestions on how to better fix this.

The patch is important, because without the arm crypto, it's pretty slow.

Before:

```
CoreELEC:~ # cryptsetup benchmark
# Tests are approximate using memory only (no storage IO).
PBKDF2-sha1       281875 iterations per second for 256-bit key
PBKDF2-sha256     569878 iterations per second for 256-bit key
PBKDF2-sha512     238312 iterations per second for 256-bit key
PBKDF2-ripemd160  152409 iterations per second for 256-bit key
PBKDF2-whirlpool     N/A
argon2i       4 iterations, 169514 memory, 4 parallel threads (CPUs) for 256-bit key (requested 2000 ms time)
argon2id      4 iterations, 186603 memory, 4 parallel threads (CPUs) for 256-bit key (requested 2000 ms time)
#     Algorithm |       Key |      Encryption |      Decryption
        aes-cbc        128b        37.0 MiB/s        38.7 MiB/s
    serpent-cbc        128b               N/A               N/A
    twofish-cbc        128b               N/A               N/A
        aes-cbc        256b        29.5 MiB/s        30.4 MiB/s
    serpent-cbc        256b               N/A               N/A
    twofish-cbc        256b               N/A               N/A
        aes-xts        256b        39.9 MiB/s        39.6 MiB/s
    serpent-xts        256b               N/A               N/A
    twofish-xts        256b               N/A               N/A
        aes-xts        512b        31.0 MiB/s        28.6 MiB/s
    serpent-xts        512b               N/A               N/A
    twofish-xts        512b               N/A               N/A
```


After:


```
CoreELEC:~ # modprobe aes_ce_blk
CoreELEC:~ # cryptsetup benchmark
# Tests are approximate using memory only (no storage IO).
PBKDF2-sha1       281875 iterations per second for 256-bit key
PBKDF2-sha256     582542 iterations per second for 256-bit key
PBKDF2-sha512     238312 iterations per second for 256-bit key
PBKDF2-ripemd160  152409 iterations per second for 256-bit key
PBKDF2-whirlpool     N/A
argon2i       4 iterations, 168870 memory, 4 parallel threads (CPUs) for 256-bit key (requested 2000 ms time)
argon2id      4 iterations, 173398 memory, 4 parallel threads (CPUs) for 256-bit key (requested 2000 ms time)
#     Algorithm |       Key |      Encryption |      Decryption
        aes-cbc        128b       282.5 MiB/s       284.2 MiB/s
    serpent-cbc        128b               N/A               N/A
    twofish-cbc        128b               N/A               N/A
        aes-cbc        256b       244.6 MiB/s       244.8 MiB/s
    serpent-cbc        256b               N/A               N/A
    twofish-cbc        256b               N/A               N/A
        aes-xts        256b       264.9 MiB/s       265.1 MiB/s
    serpent-xts        256b               N/A               N/A
    twofish-xts        256b               N/A               N/A
        aes-xts        512b       235.6 MiB/s       236.3 MiB/s
    serpent-xts        512b               N/A               N/A
    twofish-xts        512b               N/A               N/A
```