@Ray-future 2019-07-14
The Linux patch should go into the Linux branch. Other than that I would be OK with the change if it was up to me.