@afl1 2019-07-14
I think is better fix this issue, as hack code which is part of mainline linux. This issue was fix in 4.9 [kernel]() by removing _id in MODULE_GENERIC_TABLE macro. [https://github.com/CoreELEC/linux-amlogic/commit/6301939d97d079f0d3dbe71e750f4daf5d39fc33#diff-4b68ce3c700ded96468c46aa417fb441](url)