@doozan 2019-07-14
Thank you for pointing me in the right direction @afl1, that's a much better fix!

I've added packages for cryptsetup and lvm2 and put cryptsetup on the busybox dependency list.  This will put everything needed to boot with an encrypted /storage partition into the default image.

The user will need to supply his own mount-storage.sh to handle unlocking, but I'll post an example here as a proof-of-concept.  This script temporarily sets up system, loads the kernel modules, and starts SSH waiting for the user to connect and unlock the storage.

To unlock the device remotely, run `ssh root@coreelec.ip.address 'echo -n MYPASSWORD | /bin/cryptsetup luksOpen /dev/MY_LUKS_DEVICE storage-crypt -d - && killall sshd'`

/flash/mount-storage.sh
```
#!/bin/sh

echo "setting up temp SYSTEM"

mount_part "/flash/$IMAGE_SYSTEM" "/sysroot" "ro,loop"
/usr/bin/busybox mount --bind /dev /sysroot/dev
/usr/bin/busybox mount --bind /proc /sysroot/proc
/usr/bin/busybox mount --bind /sys /sysroot/sys
/usr/bin/busybox mount --bind /flash /sysroot/flash
mount -t tmpfs tmpfs /sysroot/var
mount -t tmpfs tmpfs /sysroot/run
mount -t tmpfs tmpfs /sysroot/storage

mkdir /sysroot/var/lib
chroot /sysroot /usr/bin/ln -s /usr/lib/kernel-overlays/base/lib/modules /var/lib/modules
chroot /sysroot /usr/sbin/modprobe dm-crypt
chroot /sysroot /usr/sbin/modprobe aes-ce-blk
chroot /sysroot /usr/sbin/modprobe algif_skcipher


echo "setting up sshd"

mkdir -p /sysroot/var/empty
mkdir -p /sysroot/storage/.cache/ssh/

# This is unencrypted, don't re-use the hostkey from your encrypted /storage/.ssh folder
cat<<'EOF'>/sysroot/storage/.cache/ssh/ssh_host_ed25519_key
PASTE_SSH_HOST_KEY_HERE
EOF
/sysroot/usr/bin/busybox chmod 600 /sysroot/storage/.cache/ssh/ssh_host_ed25519_key


# Uncomment this if you want to allow password login (password is corelec)
#cat<<'EOF'>/sysroot/storage/.cache/shadow
#root:$6$j2kltGrL/rskio2D$Leynsm6Dk/TBBxsdhRCNqjjwy4wnfwshN6lDnrNKMGejw9qVjlnwhQlmUYekno.oSpcDpPRxJNkUCwGJrGxLj0:::::::
#EOF
#/sysroot/usr/bin/busybox chmod 600 /sysroot/storage/.cache/shadow

mkdir /sysroot/storage/.ssh
cat<<'EOF'>/sysroot/storage/.ssh/authorized_keys
PASTE_SSH_PUBLIC_KEY_HERE
EOF
/sysroot/usr/bin/busybox chmod 600 /sysroot/storage/.ssh/authorized_keys

echo "Starting ssh"
chroot /sysroot /usr/sbin/sshd -D

# User must connect via ssh, unlock storage as storage-crypt and kill sshd

echo "Cleaning up temp system"
umount /sysroot/run
umount /sysroot/flash
umount /sysroot/dev
umount /sysroot/proc
umount /sysroot/sys
umount /sysroot/var

umount /sysroot/storage
umount /sysroot

if [ -b /dev/mapper/storage-crypt ]; then
  echo "Mounting encrypted storage..."
  mount_part "/dev/mapper/storage-crypt" "/storage" "rw,noatime"
else
  echo "storage-crypt missing, mounting default storage..."
  mount_part "$disk" "/storage" "rw,noatime"
fi
```