@doozan 2020-01-01
I'm still interesting in seeing this merged, is there anything more needed from my side?

It seems like it should be pretty straight forward to include the first patch so that we include the same kernel modules that are already being built for the Allwinner/Rockchip configs.