@adamg88 2020-01-10
Apologies that this has been left for many months, I'm good with accepting this into CE but your cryptsetup and lvm2 package.mk are not in good alignment with the rest of the buildroot packages, there is a template you can look at [here](https://raw.githubusercontent.com/CoreELEC/CoreELEC/coreelec-9.2/packages/packages.mk.template).

I would also suggest submitting the PR upstream to LE for further review.