@doozan 2020-05-13
I opened #230 with a patch to include the same kernel modules that are already being built for the Allwinner/Rockchip configs.  That's the minimum needed to make LUKS viable on CoreELEC.  If/when that gets merged, I'll work on up-streaming the other patches to build the user land binaries